+++ pyproject.toml
@@ -3,6 +3,7 @@ ignore_missing_imports = true
 disallow_incomplete_defs = true
 disallow_untyped_calls = true
 disallow_untyped_defs = true
+plugins='pydantic.mypy'
 
 [tool.isort]
 profile = "black"
