+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 disable_test_id_escaping_and_forfeit_all_rights_to_community_support = true
-addopts = -s -v --disable-pytest-warnings --cov=app --cov-report=html --cov-report term --cov-branch --cov-fail-under=60 .
+addopts = -s -v --disable-pytest-warnings --cov=app --cov-report=html --cov-report term --cov-branch --cov-fail-under=60
 pythonpath = . app
 asyncio_mode = auto
 python_files = *_test.py *_tests.py test_*.py
